Bug #1222098  cpufreq: sprd: Free memory at error handling label
[root cause  ] Need to free memory when an error occurs
[changes     ] Modify driver
[side effects] no
[self test   ]: dvfs test
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:all
[test case]:dvfs test,reboot test,sleep wakeup
[reviewers   ] jack.liu ling_ling.xu jingchao.ye
[change_type ] bringup --> sprd_adaptation
[tag_product ] common

Need to free memory when an error occurs for saving memeory

(cherry picked from commit 542edb6d2dee6a71968d4f86977da934c695f71a)

+++ drivers/cpufreq/sprd-cpufreq-common.h
@@ -85,7 +85,7 @@ struct sprd_cpufreq_driver_data {
 	unsigned int volt_tol;
 	unsigned int volt_tol_var;
 	/*BitX = 1 means sub_cluster exits*/
-	unsigned int sub_cluster_bits;
+	unsigned long sub_cluster_bits;
 	/*Volt requested by this cluster*/
 	unsigned long volt_req;/*uv*/
 	unsigned int volt_share_hosts_bits;
